+++ sdk/core/azure-core/src/http/bearer_token_authentication_policy.cpp
@@ -16,9 +16,10 @@ std::unique_ptr<RawResponse> BearerTokenAuthenticationPolicy::Send(
   {
     std::lock_guard<std::mutex> lock(m_accessTokenMutex);
 
+    // Refresh the token in 2 or less minutes before the actual expiration.
     if ((std::chrono::system_clock::now() - std::chrono::minutes(2)) > m_accessToken.ExpiresOn)
     {
-      m_accessToken = m_credential->GetToken(context, m_getTokenOptions);
+      m_accessToken = m_credential->GetToken(context, m_tokenRequestOptions);
     }
 
     request.AddHeader("authorization", "Bearer " + m_accessToken.Token);