+++ sdk/storage/azure-storage-blobs/test/page_blob_client_test.cpp
@@ -208,6 +208,7 @@ namespace Azure { namespace Storage { namespace Test {
     EXPECT_TRUE(cLease.ETag.HasValue());
     EXPECT_TRUE(IsValidTime(cLease.LastModified));
     EXPECT_EQ(cLease.LeaseId, leaseId2);
+    leaseClient = Blobs::BlobLeaseClient(*m_pageBlobClient, cLease.LeaseId);
     EXPECT_EQ(leaseClient.GetLeaseId(), leaseId2);
 
     auto blobInfo = *leaseClient.Release();