+++ sdk/storage/azure-storage-files-shares/test/share_file_client_test.cpp
@@ -272,6 +272,8 @@ namespace Azure { namespace Storage { namespace Test {
     EXPECT_TRUE(cLease.ETag.HasValue());
     EXPECT_TRUE(cLease.LastModified >= lastModified);
     EXPECT_EQ(cLease.LeaseId, leaseId2);
+    leaseClient = Files::Shares::ShareLeaseClient(*m_fileClient, cLease.LeaseId);
+    EXPECT_EQ(leaseClient.GetLeaseId(), leaseId2);
 
     lastModified = m_fileClient->GetProperties()->LastModified;
     auto fileInfo = *leaseClient.Release();